requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.0

# Web Search APIs
tavily-python>=0.7.0
exa-py>=1.0.0

# Notion Integration
notion-client>=2.2.0

# HTTP Client
httpx>=0.27.0

# Environment Management
python-dotenv>=1.0.0

# Async Support
anyio>=4.0.0

# Data Validation
pydantic>=2.0.0

# Browser Automation
playwright>=1.40.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
